# libs/data-access/src/lib/models/listing-draft-created-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface ListingDraftCreatedDto {
  email: string;
  id: string;
  language: 'ro' | 'en';

  /**
   * Whether a link e-mail was queued
   */
  linkSent?: boolean;

  /**
   * When no link was sent: seconds until one can be
   */
  retryAfterSeconds?: number;
  status: 'open' | 'submitted';
  step: number;

  /**
   * This browser's key to the draft
   */
  token?: string;
  updatedAt: string;
}